feat(NodeAuthoring): Show component preview in current language
