+++ src/assets/wise5/services/studentProjectTranslationService.ts
@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
-import { lastValueFrom, of, switchMap, tap } from 'rxjs';
+import { lastValueFrom, of, tap } from 'rxjs';
 import { copy } from '../common/object/object';
 import { Translations } from '../../../app/domain/translations';
 import { ProjectTranslationService } from './projectTranslationService';
-import { toSignal, toObservable } from '@angular/core/rxjs-interop';
 import { Language } from '../../../app/domain/language';
 import { HttpClient } from '@angular/common/http';
 import { ConfigService } from './configService';
@@ -20,17 +19,6 @@ export class StudentProjectTranslationService extends ProjectTranslationService 
   ) {
     super(configService, http, projectService);
   }
-
-  currentTranslations = toSignal(
-    toObservable(this.projectService.currentLanguage).pipe(
-      switchMap((language) =>
-        this.projectService.getLocale().isDefaultLocale(language.locale)
-          ? of({})
-          : lastValueFrom(this.fetchTranslations(language.locale))
-      )
-    ),
-    { initialValue: {} }
-  );
 
   async switchLanguage(language: Language, requester: 'student' | 'system'): Promise<void> {
     this.projectService.setCurrentLanguage(language);
@@ -63,24 +51,5 @@ export class StudentProjectTranslationService extends ProjectTranslationService 
     const project = copy(this.projectService.getOriginalProject());
     this.projectService.setProject(project);
     return project;
-  }
-
-  private applyTranslations(projectElement: object, translations: Translations): void {
-    Object.keys(projectElement)
-      .filter((key) => key.endsWith('.i18n'))
-      .forEach((key) => {
-        const translationKey = projectElement[key].id;
-        if (translations[translationKey]) {
-          const keyWithoutI18NId = key.substring(0, key.lastIndexOf('.i18n'));
-          projectElement[keyWithoutI18NId] = translations[translationKey].value;
-        }
-      });
-    Object.values(projectElement).forEach((value) => {
-      if (Array.isArray(value)) {
-        value.forEach((val) => this.applyTranslations(val, translations));
-      } else if (typeof value === 'object' && value != null) {
-        this.applyTranslations(value, translations);
-      }
-    });
   }
 }
