+++ src/messages.xlf
@@ -10649,7 +10649,7 @@ Click &quot;Cancel&quot; to keep the invalid JSON open so you can fix it.</sourc
         <source>Edit content</source>
         <context-group purpose="location">
           <context context-type="sourcefile">src/assets/wise5/authoringTool/components/component-authoring.component.ts</context>
-          <context context-type="linenumber">44,49</context>
+          <context context-type="linenumber">46,51</context>
         </context-group>
       </trans-unit>
       <trans-unit id="4602079498884979331" datatype="html">
